# perf(scip-syntax): uses the same parse tree for globals and locals