--- a/docker-images/syntax-highlighter/crates/scip-syntax/Cargo.toml
+++ b/docker-images/syntax-highlighter/crates/scip-syntax/Cargo.toml
@@ -25,6 +25,7 @@ string-interner = { workspace = true }
 walkdir = { workspace = true }
 path-clean = { workspace = true }
 camino = { workspace = true }
+tree-sitter = { workspace = true }
 
 syntax-analysis = { path = "../syntax-analysis" }
 tree-sitter-all-languages = { path = "../tree-sitter-all-languages" }