… docker-images/syntax-highlighter/Cargo.lock