Closes https://linear.app/sourcegraph/issue/GRAPH-718/dont-parse-every-file-twice-in-scip-syntax

While working on parallel scip-syntax I noticed we're re-parsing every file for both globals and locals generation. This PR makes it so we use the same parse tree for both. I've also made it so we don't re-initialize the language parsers for every file. The thread-local-storage aspect of this matters after merging #63536 

## Test plan

```
spring-framework on main [?]
❯ hyperfine --warmup 1 'git archive HEAD | scip-syntax-main index tar - -l java -o main.scip' 'git archive HEAD | scip-syntax index tar - -l java -o single-parse.scip'
Benchmark 1: git archive HEAD | scip-syntax-main index tar - -l java -o main.scip
  Time (mean ± σ):      6.894 s ±  0.033 s    [User: 6.877 s, System: 0.093 s]
  Range (min … max):    6.814 s …  6.921 s    10 runs

Benchmark 2: git archive HEAD | scip-syntax index tar - -l java -o single-parse.scip
  Time (mean ± σ):      4.675 s ±  0.008 s    [User: 4.693 s, System: 0.084 s]
  Range (min … max):    4.663 s …  4.688 s    10 runs

Summary
  git archive HEAD | scip-syntax index tar - -l java -o single-parse.scip ran
    1.47 ± 0.01 times faster than git archive HEAD | scip-syntax-main index tar - -l java -o main.scip

spring-framework on  main [?]
❯ scip-syntax scip-evaluate --ground-truth main.scip --candidate single-parse.scip
{"precision_percent":"100.0","recall_percent":"100.0","true_positives":"569982.0","false_positives":"0.0","false_negatives":"0.0"}
```